fix int64 casts with NTL

+++ linbox/ring/ntl/ntl-lzz_pe.h
@@ -402,7 +402,7 @@ namespace LinBox
                 if(_seed == 0)
                     NTL::SetSeed(NTL::to_ZZ(time(0)));
                 else
-                    NTL::SetSeed(NTL::to_ZZ(_seed));
+                    NTL::SetSeed(NTL::to_ZZ(static_cast<long unsigned int>(_seed)));
             }
 
         const NTL_zz_pE& ring() const { return _ring; }
@@ -413,7 +413,7 @@ namespace LinBox
                 if(_seed == 0)
                     NTL::SetSeed(NTL::to_ZZ(time(0)));
                 else
-                    NTL::SetSeed(NTL::to_ZZ(_seed));
+                    NTL::SetSeed(NTL::to_ZZ(static_cast<long unsigned int>(_seed)));
             }
 
 		Element& random (Element& x) const
